# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AEROVAULTT Attendance API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type Role = typeof Role[keyof typeof Role];


export const Role = {
  employee: 'employee',
  hr: 'hr',
  admin: 'admin',
} as const;

export type AttendanceStatus = typeof AttendanceStatus[keyof typeof AttendanceStatus];


export const AttendanceStatus = {
  present: 'present',
  late: 'late',
  half_day: 'half_day',
  absent: 'absent',
  on_leave: 'on_leave',
} as const;

export type LeaveType = typeof LeaveType[keyof typeof LeaveType];


export const LeaveType = {
  casual: 'casual',
  sick: 'sick',
  permission: 'permission',
  wfh: 'wfh',
} as const;

export type LeaveStatus = typeof LeaveStatus[keyof typeof LeaveStatus];


export const LeaveStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface LoginInput {
  employeeId: string;
  password: string;
}

export interface ChangePasswordInput {
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface ResetPasswordInput {
  /** @minLength 6 */
  newPassword: string;
}

export interface DeactivateInput {
  isActive: boolean;
}

export interface Employee {
  id: number;
  employeeId: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: Role;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  departmentName?: string | null;
  /** @nullable */
  officeLocationId?: number | null;
  /** @nullable */
  officeLocationName?: string | null;
  /** @nullable */
  shiftStart?: string | null;
  /** @nullable */
  shiftEnd?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface EmployeeInput {
  employeeId: string;
  name: string;
  email: string;
  phone?: string;
  /** @minLength 6 */
  password: string;
  role: Role;
  departmentId?: number;
  officeLocationId?: number;
  shiftStart?: string;
  shiftEnd?: string;
}

export interface EmployeeUpdate {
  name?: string;
  email?: string;
  phone?: string;
  role?: Role;
  departmentId?: number;
  officeLocationId?: number;
  shiftStart?: string;
  shiftEnd?: string;
}

export interface AuthResponse {
  token: string;
  user: Employee;
}

export interface Department {
  id: number;
  name: string;
  employeeCount?: number;
}

export interface DepartmentInput {
  /** @minLength 1 */
  name: string;
}

export interface OfficeLocation {
  id: number;
  name: string;
  latitude: number;
  longitude: number;
  radiusMeters: number;
}

export interface OfficeLocationInput {
  /** @minLength 1 */
  name: string;
  latitude: number;
  longitude: number;
  /** @minimum 1 */
  radiusMeters: number;
}

export interface Attendance {
  id: number;
  employeeId: number;
  employeeName: string;
  /** @nullable */
  departmentName?: string | null;
  date: string;
  /** @nullable */
  checkInTime?: string | null;
  /** @nullable */
  checkInLatitude?: number | null;
  /** @nullable */
  checkInLongitude?: number | null;
  /** @nullable */
  checkInAccuracy?: number | null;
  /** @nullable */
  checkOutTime?: string | null;
  /** @nullable */
  checkOutLatitude?: number | null;
  /** @nullable */
  checkOutLongitude?: number | null;
  /** @nullable */
  checkOutAccuracy?: number | null;
  /** @nullable */
  workingHours?: number | null;
  status: AttendanceStatus;
  locationVerified: boolean;
  /** @nullable */
  deviceInfo?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface CheckInInput {
  latitude: number;
  longitude: number;
  accuracy: number;
  deviceInfo?: string;
}

export interface CheckOutInput {
  latitude: number;
  longitude: number;
  accuracy: number;
}

export interface AttendanceCorrection {
  checkInTime?: string;
  checkOutTime?: string;
  status?: AttendanceStatus;
  notes?: string;
}

export interface Leave {
  id: number;
  employeeId: number;
  employeeName: string;
  type: LeaveType;
  startDate: string;
  endDate: string;
  reason: string;
  /** @nullable */
  attachmentUrl?: string | null;
  status: LeaveStatus;
  /** @nullable */
  reviewNotes?: string | null;
  createdAt: string;
}

export interface LeaveInput {
  type: LeaveType;
  startDate: string;
  endDate: string;
  /** @minLength 1 */
  reason: string;
  attachmentUrl?: string;
}

export interface LeaveReview {
  status: LeaveStatus;
  reviewNotes?: string;
}

export interface Holiday {
  id: number;
  name: string;
  date: string;
}

export interface HolidayInput {
  /** @minLength 1 */
  name: string;
  date: string;
}

export interface Notification {
  id: number;
  message: string;
  type: string;
  isRead: boolean;
  createdAt: string;
}

export interface EmployeeDashboardSummary {
  status: AttendanceStatus;
  /** @nullable */
  checkInTime: string | null;
  /** @nullable */
  checkOutTime: string | null;
  /** @nullable */
  workingHoursToday: number | null;
  officeVerified: boolean;
}

export interface HrDashboardSummary {
  presentToday: number;
  absentToday: number;
  lateArrivals: number;
  pendingLeaveRequests: number;
  checkedIn: number;
  yetToCheckIn: number;
  attendancePercentage: number;
}

export type ListEmployeesParams = {
departmentId?: number;
isActive?: boolean;
};

export type GetAttendanceHistoryParams = {
month?: number;
year?: number;
};

export type ListAttendanceParams = {
date?: string;
departmentId?: number;
status?: AttendanceStatus;
employeeId?: number;
};

export type ListLeavesParams = {
status?: LeaveStatus;
employeeId?: number;
};