# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AEROVAULTT Attendance API specification
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with employee ID and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "200":
          description: Updated department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /attendance/check-in:
    post:
      operationId: checkIn
      tags: [attendance]
      summary: Check in for today (requires geofence verification)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInInput"
      responses:
        "201":
          description: Checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
        "400":
          description: Outside geofence, poor accuracy, or already checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/check-out:
    post:
      operationId: checkOut
      tags: [attendance]
      summary: Check out for today
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckOutInput"
      responses:
        "200":
          description: Checked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
        "400":
          description: No active check-in found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/history:
    get:
      operationId: getAttendanceHistory
      tags: [attendance]
      summary: Get the current employee's attendance history
      parameters:
        - name: month
          in: query
          schema: { type: integer }
        - name: year
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Attendance history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List all attendance records (HR/Admin)
      parameters:
        - name: date
          in: query
          schema: { type: string, format: date }
        - name: departmentId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/AttendanceStatus"
        - name: employeeId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
  /attendance/{id}:
    patch:
      operationId: correctAttendance
      tags: [attendance]
      summary: Manually correct an attendance record (HR/Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceCorrection"
      responses:
        "200":
          description: Updated attendance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
    Role:
      type: string
      enum: [employee, hr, admin]
    AttendanceStatus:
      type: string
      enum: [present, late, half_day, absent, on_leave]
...
    LoginInput:
      type: object
      required: [employeeId, password]
      properties:
        employeeId: { type: string }
        password: { type: string }
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string, minLength: 6 }
...
    Employee:
      type: object
      required:
        - id
        - employeeId
        - name
        - email
        - role
        - isActive
        - createdAt
      properties:
        id: { type: integer }
        employeeId: { type: string }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        role:
          $ref: "#/components/schemas/Role"
        departmentId: { type: ["integer", "null"] }
        departmentName: { type: ["string", "null"] }
        officeLocationId: { type: ["integer", "null"] }
        officeLocationName: { type: ["string", "null"] }
        shiftStart: { type: ["string", "null"] }
        shiftEnd: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/Employee"
    Department:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
        employeeCount: { type: integer }
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
...
    Attendance:
      type: object
      required:
        - id
        - employeeId
        - employeeName
        - date
        - status
        - locationVerified
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        employeeName: { type: string }
        departmentName: { type: ["string", "null"] }
        date: { type: string, format: date }
        checkInTime: { type: ["string", "null"], format: date-time }
        checkInLatitude: { type: ["number", "null"] }
        checkInLongitude: { type: ["number", "null"] }
        checkInAccuracy: { type: ["number", "null"] }
        checkOutTime: { type: ["string", "null"], format: date-time }
        checkOutLatitude: { type: ["number", "null"] }
        checkOutLongitude: { type: ["number", "null"] }
        checkOutAccuracy: { type: ["number", "null"] }
        workingHours: { type: ["number", "null"] }
        status:
          $ref: "#/components/schemas/AttendanceStatus"
        locationVerified: { type: boolean }
        deviceInfo: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
    CheckInInput:
      type: object
      required: [latitude, longitude, accuracy]
      properties:
        latitude: { type: number }
        longitude: { type: number }
        accuracy: { type: number }
        deviceInfo: { type: string }
    CheckOutInput:
      type: object
      required: [latitude, longitude, accuracy]
      properties:
        latitude: { type: number }
        longitude: { type: number }
        accuracy: { type: number }
    AttendanceCorrection:
      type: object
      properties:
        checkInTime: { type: string, format: date-time }
        checkOutTime: { type: string, format: date-time }
        status:
          $ref: "#/components/schemas/AttendanceStatus"
        notes: { type: string }
...