# lib/api-zod/src/generated/types/attendanceCorrection.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AEROVAULTT Attendance API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceStatus } from './attendanceStatus';

export interface AttendanceCorrection {
  checkInTime?: Date;
  checkOutTime?: Date;
  status?: AttendanceStatus;
  notes?: string;
}