# lib/api-zod/src/generated/types/attendanceStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AEROVAULTT Attendance API specification
 * OpenAPI spec version: 0.1.0
 */

export type AttendanceStatus = typeof AttendanceStatus[keyof typeof AttendanceStatus];


export const AttendanceStatus = {
  present: 'present',
  late: 'late',
  half_day: 'half_day',
  absent: 'absent',
  on_leave: 'on_leave',
} as const;