# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AEROVAULTT Attendance API specification
 * OpenAPI spec version: 0.1.0
 */

export type Role = typeof Role[keyof typeof Role];


export const Role = {
  employee: 'employee',
  hr: 'hr',
  admin: 'admin',
} as const;